requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
mysqlclient==2.2.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.10.0
openpyxl==3.1.2
phonenumbers==8.13.30
reportlab==4.1.0
twilio==9.0.5
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
bcrypt==4.1.2
